requirements.txt
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
awscli==1.33.26
babel==2.16.0
beautifulsoup4==4.12.3
behave==1.2.6
bleach==6.2.0
boto3==1.34.144
botocore==1.34.144
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cssselect==1.2.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
docutils==0.16
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
fake-useragent==1.5.1
fasteners==0.19
fastjsonschema==2.20.0
filelock==3.16.1
fqdn==1.5.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==1.26.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
parso==0.8.4
pdbp==1.5.4
platformdirs==4.3.6
playwright==1.43.0
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.0
pycparser==2.22
pyee==11.1.0
Pygments==2.18.0
pynose==1.5.3
pyotp==2.9.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.3
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.20.1
rsa==4.7.2
s3transfer==0.10.2
sbvirtualdisplay==1.3.0
selenium==4.25.0
seleniumbase==4.31.6
Send2Trash==1.8.3
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tabcompleter==1.3.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
